# openapi/arpa-openapi.yaml
openapi: 3.1.0
info:
  title: ARPA Protocol API
  version: 0.9.4
  description: Machine-readable contract for ARPA Core and optional modules. Registration,
    publication and discovery do not imply authority, assurance or trust.
servers:
- url: http://127.0.0.1:8000
paths:
  /health:
    get:
      operationId: health
      responses:
        '200':
          description: Healthy
  /registry:
    get:
      operationId: getRegistryMetadata
      responses:
        '200':
          description: Registry metadata
          content:
            application/json:
              schema:
                $ref: ../schemas/registry-metadata.schema.json
  /agents:
    post:
      operationId: registerAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/agent-core.schema.json
      responses:
        '201':
          description: Registered
        '409':
          $ref: '#/components/responses/Problem'
    get:
      operationId: listAgents
      description: Return caller-visible agent publication summaries. Discoverability
        does not imply authority or permission to invoke.
      parameters:
      - name: skill
        in: query
        schema:
          type: string
      - name: capability
        in: query
        schema:
          type: string
      - name: protocolVersion
        in: query
        schema:
          type: string
      - name: namespace
        in: query
        schema:
          type: string
      - name: disclosureClass
        in: query
        schema:
          type: string
          enum:
          - public
          - authenticated-extended
          - tenant-specific
          - private
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 200
          default: 50
      - name: cursor
        in: query
        schema:
          type: string
      responses:
        '200':
          description: Caller-visible publication summaries
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                - next_cursor
                properties:
                  items:
                    type: array
                    items:
                      $ref: ../schemas/a2a-publication-projection.schema.json
                  next_cursor:
                    type:
                    - string
                    - 'null'
  /agents/{agent_id}:
    parameters:
    - $ref: '#/components/parameters/AgentId'
    - $ref: '#/components/parameters/At'
    get:
      operationId: resolveAgent
      responses:
        '200':
          description: Current or point-in-time agent record. Historical callers may use the dedicated resolution endpoint for explicit requested-time/current-state provenance.
        '404':
          $ref: '#/components/responses/Problem'
  /agents/{agent_id}/historical-resolution:
    parameters:
    - $ref: '#/components/parameters/AgentId'
    - name: at
      in: query
      required: true
      schema:
        type: string
        format: date-time
    get:
      operationId: resolveHistoricalAuthority
      responses:
        '200':
          description: Deterministic historical authority resolution separating requested-time and current state
          content:
            application/json:
              schema:
                $ref: ../schemas/historical-resolution.schema.json
        '404':
          $ref: '#/components/responses/Problem'

  /agents/{agent_id}/status:
    parameters:
    - $ref: '#/components/parameters/AgentId'
    - $ref: '#/components/parameters/At'
    get:
      operationId: getAgentStatus
      responses:
        '200':
          description: Multidimensional status
          content:
            application/json:
              schema:
                $ref: ../schemas/status.schema.json
  /agents/{agent_id}/events:
    parameters:
    - $ref: '#/components/parameters/AgentId'
    - name: after
      in: query
      schema:
        type: integer
        minimum: 0
    get:
      operationId: listAgentEvents
      responses:
        '200':
          description: Ordered events
  /aliases/{alias}:
    parameters:
    - name: alias
      in: path
      required: true
      schema:
        type: string
    - $ref: '#/components/parameters/At'
    get:
      operationId: resolveAlias
      responses:
        '200':
          description: Alias resolution
        '409':
          $ref: '#/components/responses/Problem'
  /records/{record_id}:
    parameters:
    - name: record_id
      in: path
      required: true
      schema:
        type: string
    get:
      operationId: getRecord
      responses:
        '200':
          description: Record
        '404':
          $ref: '#/components/responses/Problem'
  /authority/evaluate:
    post:
      operationId: evaluateAuthority
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Authority decision and decision receipt
        '501':
          $ref: '#/components/responses/Problem'
  /reliance/evaluate:
    post:
      operationId: evaluateReliance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Reliance decision
        '501':
          $ref: '#/components/responses/Problem'
  /events:
    get:
      operationId: listEvents
      parameters:
      - name: after
        in: query
        schema:
          type: integer
          minimum: 0
      responses:
        '200':
          description: Replayable event stream
components:
  parameters:
    AgentId:
      name: agent_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^agentreg:'
    At:
      name: at
      in: query
      required: false
      schema:
        type: string
        format: date-time
  schemas:
    Problem:
      type: object
      required:
      - type
      - title
      - status
      - code
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        detail:
          type: string
        instance:
          type: string
        correlation_id:
          type: string
        reason_codes:
          type: array
          items:
            type: string
        retryable:
          type: boolean
  responses:
    Problem:
      description: ARPA problem response
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'