# reference/app.py
from __future__ import annotations
import hashlib, json, os, uuid
from datetime import datetime, timezone
from fastapi import FastAPI, HTTPException, Query
from fastapi.responses import JSONResponse
from .storage import Store
from scripts.reference_evaluator import evaluate_authority

app=FastAPI(title='ARPA Reference Service',version='0.9.4')
store=Store(os.getenv('ARPA_DB',':memory:'))

def now(): return datetime.now(timezone.utc).isoformat().replace('+00:00','Z')
def problem(status,code,title,detail='',reasons=None,retryable=False):
    return JSONResponse(status_code=status,content={'type':f'https://arpa.example/problems/{code.lower()}','title':title,'status':status,'code':code,'detail':detail,'correlation_id':str(uuid.uuid4()),'reason_codes':reasons or [],'retryable':retryable})
def emit(subject,event_type,payload=None):
    event={'event_id':str(uuid.uuid4()),'event_type':event_type,'subject':subject,'source':'registry:reference','effective_at':now(),'issued_at':now(),'payload':payload or {}}
    event['sequence']=store.add_event(event); return event

@app.get('/health')
def health(): return {'status':'ok','version':'0.9.4'}

@app.get('/registry')
def registry():
    return {'registry_id':os.getenv('ARPA_REGISTRY_ID','registry:reference'),'name':os.getenv('ARPA_REGISTRY_NAME','ARPA Reference Registry'),'arpa_version':'0.9.0','implementation_release':'0.9.4','supported_modules':['ARPA-Core','ARPA-Relations','ARPA-Authority','ARPA-Evidence'],'supported_profiles':['A','C'],'authoritative_base_uri':os.getenv('ARPA_PUBLIC_BASE_URI','http://127.0.0.1:8000'),'conformance_declaration_uri':'http://127.0.0.1:8000/records/conformance-reference','event_retention_seconds':86400,'status_max_age_seconds':300}


@app.get('/agents')
def list_agents(skill:str|None=None, capability:str|None=None, protocolVersion:str|None=None, namespace:str|None=None, disclosureClass:str|None=None, limit:int=Query(50,ge=1,le=200), cursor:str|None=None):
    records=store.all_records()
    cores={r.get('agent_id') or r.get('subject'):r for r in records if r.get('record_type')=='agent_core'}
    descriptions={}
    for r in records:
        if r.get('record_type')=='agent-description-reference':
            aid=r.get('agent_id') or r.get('subject')
            descriptions.setdefault(aid,[]).append(r)
    items=[]
    for aid,core in sorted(cores.items()):
        refs=descriptions.get(aid,[])
        ref=refs[-1] if refs else None
        disclosure=(ref or {}).get('disclosure_class','public')
        if disclosureClass and disclosure!=disclosureClass: continue
        if disclosure!='public': continue  # reference service has no caller auth context; fail closed
        versions=(ref or {}).get('protocol_versions',[])
        if protocolVersion and protocolVersion not in versions: continue
        item={
            'publication_id':(ref or {}).get('record_id',f'publication:{aid}'),
            'agent_id':aid,
            'agent_card_uri':(ref or {}).get('uri','https://invalid.example/agent-card'),
            'card_digest':(ref or {}).get('digest','0'*64),
            'representation_version':(ref or {}).get('representation_version'),
            'snapshot_uri':None,
            'observed_at':(ref or {}).get('retrieved_at') or core.get('issued_at') or now(),
            'valid_until':(ref or {}).get('valid_until'),
            'disclosure_class':disclosure,
            'namespace':namespace,
            'lifecycle_status':core.get('status','active'),
            'publisher':core.get('issuer'),
            'protocol_versions':versions,
            'skills':[],
            'capabilities':[],
            'authority_implication':False,
            'source_record_ids':[core.get('record_id')] + ([ref.get('record_id')] if ref else [])
        }
        if skill and skill not in item['skills']: continue
        if capability and capability not in item['capabilities']: continue
        items.append(item)
    return {'items':items[:limit],'next_cursor':None}

@app.post('/agents',status_code=201)
def register_agent(record:dict):
    aid=record.get('agent_id') or record.get('subject')
    if not isinstance(aid,str) or not aid.startswith('agentreg:'): raise HTTPException(422,'agent_id must use agentreg scheme')
    existing=store.records_for_subject(aid)
    if any(r.get('record_type')=='agent_core' for r in existing): return problem(409,'ARPA-STATE-TRANSITION-INVALID','Identifier already issued','Canonical identifiers are non-reassignable')
    store.put_record(record); emit(aid,'agent.registered',{'record_id':record['record_id']}); return record

@app.post('/records',status_code=201)
def put_record(record:dict):
    store.put_record(record)
    subject=record.get('agent_id') or record.get('subject')
    if record.get('record_type')=='identifier-alias': store.add_alias(record); emit(subject,'alias.added',{'alias':record['alias']})
    else: emit(subject if isinstance(subject,str) else record['record_id'],'agent.metadata.updated',{'record_id':record['record_id'],'record_type':record['record_type']})
    return record

@app.get('/records/{record_id}')
def get_record(record_id:str):
    r=store.get_record(record_id)
    if not r: return problem(404,'ARPA-ID-NOT-FOUND','Record not found')
    return r

@app.get('/agents/{agent_id}')
def resolve_agent(agent_id:str,at:str|None=Query(None)):
    records=store.records_for_subject(agent_id,at)
    core=[r for r in records if r.get('record_type')=='agent_core']
    if not core: return problem(404,'ARPA-ID-NOT-FOUND','Agent not found','Identifier was never issued or is not visible')
    return {'agent':core[-1],'records':records,'resolution_time':now(),'at':at,'authoritativeness':'authoritative','projection_lag_seconds':0}

@app.get('/agents/{agent_id}/historical-resolution')
def historical_resolution(agent_id:str, at:str=Query(...)):
    historical=store.records_for_subject(agent_id,at)
    current=store.records_for_subject(agent_id)
    historical_status=[r for r in historical if r.get('record_type')=='status']
    current_status=[r for r in current if r.get('record_type')=='status']
    if not historical_status:
        return problem(409,'ARPA-HISTORICAL-EVIDENCE-UNAVAILABLE','Historical status unavailable','No retained status record covers the requested time', ['ARPA-HISTORICAL-RETENTION-BOUNDARY'])
    if not current_status:
        return problem(409,'ARPA-AUTHORITY-INDETERMINATE','Current status unavailable',retryable=True)
    hs=historical_status[-1]; cs=current_status[-1]
    later=[e for e in store.events(agent_id,0) if e.get('effective_at','')>at]
    material=[]
    for e in later:
        et=e.get('event_type','status_change')
        effect='prospective' if et in {'delegation.revoked','agent.suspended','agent.retired','agent.superseded','recognition.withdrawn'} else 'governance_defined'
        payload=e.get('payload') or {}
        effect=payload.get('historical_effect',effect)
        material.append({'event_type':et,'effective_at':e.get('effective_at'),'historical_effect':effect,'reason_code':payload.get('reason_code'),'evidence_reference':f"event:{e.get('sequence')}" if e.get('sequence') else None})
    selected=[]
    for r in historical:
        canonical=json.dumps(r,sort_keys=True,separators=(',',':')).encode()
        selected.append({'record_id':r.get('record_id','unknown'),'version':str(r.get('schema_version','1')),'effective_from':r.get('effective_from',r.get('issued_at')),'effective_until':r.get('effective_until'),'digest':'sha256:'+hashlib.sha256(canonical).hexdigest()})
    checkpoint=str(max([e.get('sequence',0) for e in store.events(agent_id,0)] or [0]))
    return {
        'subject':agent_id,
        'requested_time':at,
        'evaluation_time':now(),
        'state_at_requested_time':hs,
        'current_state':cs,
        'reconstruction_status':'authoritative_complete',
        'selected_records':selected,
        'event_checkpoint':'event:'+checkpoint,
        'later_material_events':material,
        'historical_effect':'governance_defined' if any(x['historical_effect']=='governance_defined' for x in material) else ('prospective' if material else 'none'),
        'retention':{'evidence_available':True,'status':'available','boundary':None},
        'evidence':{'references':['event:'+checkpoint],'integrity_status':'verified','lineage_mechanism':'repository-sqlite-event-sequence'},
        'warnings':['current state differs from historical state'] if hs!=cs else []
    }

@app.get('/agents/{agent_id}/status')
def get_status(agent_id:str,at:str|None=Query(None)):
    records=store.records_for_subject(agent_id,at)
    statuses=[r for r in records if r.get('record_type')=='status']
    if not statuses: return problem(409,'ARPA-AUTHORITY-INDETERMINATE','Status unavailable',retryable=True)
    return statuses[-1]

@app.get('/agents/{agent_id}/events')
def agent_events(agent_id:str,after:int=0): return {'events':store.events(agent_id,after),'after':after}
@app.get('/events')
def events(after:int=0): return {'events':store.events(None,after),'after':after}

@app.get('/aliases/{alias:path}')
def resolve_alias(alias:str,at:str|None=Query(None)):
    instant=at or now(); matches=store.resolve_alias(alias,instant)
    if not matches: return problem(404,'ARPA-ID-NOT-FOUND','Alias not found')
    ids=sorted({r['canonical_id'] for r in matches})
    if len(ids)>1: return problem(409,'ARPA-ID-CONFLICT','Alias conflict','Multiple active canonical identifiers', ['alias_conflict'])
    return {'alias':alias,'canonical_id':ids[0],'records':matches,'at':instant}

@app.post('/authority/evaluate')
def authority_evaluate(payload:dict):
    decision,reasons=evaluate_authority(payload)
    receipt={'record_id':str(uuid.uuid4()),'record_type':'decision-receipt','schema_version':'1.0.0','issuer':'pdp:reference','subject':payload.get('request',{}).get('agent','unknown'),'issued_at':now(),'effective_from':now(),'effective_until':None,'status':'issued','request_digest':'sha256:'+hashlib.sha256(json.dumps(payload,sort_keys=True,separators=(',',':')).encode()).hexdigest(),'decision':decision,'reason_codes':reasons,'evaluator':'pdp:reference','policy_version':'reference-0.9.4'}
    store.put_record(receipt)
    return {'decision':decision,'reason_codes':reasons,'decision_receipt':receipt}

@app.post('/reliance/evaluate')
def reliance_evaluate(payload:dict):
    required=payload.get('required_modules',[]); supported=set(registry()['supported_modules']); missing=[m for m in required if m not in supported]
    if missing: return {'decision':'indeterminate','reason_codes':['required_module_not_supported'],'missing_modules':missing}
    return {'decision':'allow','reason_codes':['required_modules_supported']}

@app.get('/interoperability')
def interoperability():
    return {
        'arpa_version':'0.9.0','implementation_release':'0.9.4',
        'exchange_profile':'arpa-interoperability-0.5',
        'supports':['registry-metadata','canonical-resolution','scoped-recognition','event-replay','revocation-acknowledgement'],
        'evidence_format':'artifacts/interoperability/evidence-bundle.json',
        'independent_implementation':False
    }